Data_Analyst_Agent/PoC/requirements.txt
# Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Data & ML
pandas>=2.2.0
numpy>=1.24.0
openai>=1.12.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Frontend
streamlit>=1.32.0
requests>=2.31.0
pillow>=10.0.0

# Utilities
watchdog>=3.0.0 